requirements.txt
asgiref==3.2.10
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytest-django==4.4.0
pytz==2020.1
sqlparse==0.3.1
